asignacion de componentes
